Move node and server contexts from CLI to common
This will allow us to drop CLI dependencies from the node and server.
