+++ vantage6-server/vantage6/server/cli/server.py
@@ -1,5 +1,6 @@
 import click
 import IPython
+from vantage6.common.globals import DEFAULT_SERVER_SYSTEM_FOLDERS as S_FOL
 from vantage6.common.globals import InstanceType
 import yaml
 
@@ -11,11 +12,10 @@ from vantage6.common import (
     info,
     error,
 )
+from vantage6.common.server_context import ServerContext
 from vantage6.server.model.base import Database
-from vantage6.cli.globals import DEFAULT_SERVER_SYSTEM_FOLDERS as S_FOL
 from vantage6.server.controller import fixture
 from vantage6.cli.configuration_wizard import select_configuration_questionaire
-from vantage6.cli.context.server import ServerContext
 from vantage6.server._version import __version__
 
 help_ = {
